requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
qrcode[pil]==7.4.2
pandas==2.1.3
openpyxl==3.1.2
gunicorn==21.2.0
Werkzeug==3.0.1
